backend/requirements.txt
fastapi[all]
uvicorn[standard]
openai
ipython
requests
bs4
python-dotenv